requirements.txt
python-dotenv==1.2.2
python-boto3==1.28.57